Adds gurdian and updates onboarded.txt format
Now onboarded.txt is a proper CSV file, with two fields:
primary_domain,sd_rewrite_rule
The file also has the updated entry for The Gurdian.

This enables us to allow the organizations to request the human
readable name (in the sd_rewrite_rule), and add that as the ruleset.