Split generate and sign steps
One of the biggest frustrations I have with this process is that the
generation and signing steps are combined. I really want to generate
everything, review it, and then move ahead with signing.

This splits the two steps and inlines them into the Makefile as separate
`make generate` and `make sign`. Also change the default goal to `help`,
which is what all of our other Makefiles do.

This necessitates checking in the default.rulesets file that's used as
the input for signing.