Disallow new onion names with subdomains
Using an extra subdomain causes issues in Tor Browser, because the
first-party domain is incorrectly determined.

Tor has asked us (in #219) to stop adding new domains, which is simple
enough. A new check verifies there's no period in the part before
".securedrop.tor.onion". If there is and it's not in the exemption list,
error out.

Refs #219.