Add second onion name for abc.au
To facilitate changing the short onion name for abc.au to no longer use
a subdomain (as requested by Tor in #219), we're setting up a second
name for them, and once they've migrated over, will remove the original
one.

A hack is added to the generation script to save the old domain under
"abc-legacy" and keep the new domain as just "abc".

Refs #222.